requirements.txt
python-telegram-bot==20.7
python-telegram-bot[socks]==20.7
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.1.1